Build all runtime requirements before build
I'm not 100% sure if it is the "correct" way or not, but some dists only
specify build-specific deps, and assume that runtime deps are present
before building.  This is in contrast to most packaging solutions like
RPM or DPKG, which treat the two lists as separate and independent.

For now, this makes Dancer::Plugin::DBIC build properly, without
intervention or supervision.