Generate files list outside of build directory
Some File::Find::* modules do find() operations inside their build root,
and get very upset when the filelist metafile exists there.  So, instead
we'll just create the metafile in %_tmppath, and sidestep the whole
problem.

Made the filelist a macro so that naming can be adjusted later with
minimal changes to the SPEC file template.